src/requirements.txt
numpy
Pillow
torch
onnxruntime
